+++ scripts/changelog.ts
@@ -299,7 +299,8 @@ async function exec(): Promise<void> {
       ]
     },
     {
-      changed: ['Optimize bundle size']
+      changed: ['Optimize bundle size'],
+      added: ['**plugin-table**. You can now create tables in markdown.']
     }
   ])