+++ packages/bundle-size/src/plugin-video.tsx
@@ -1,0 +1,1 @@
+export { videoPlugin } from '@edtr-io/plugin-video'