+++ scripts/changelog.ts
@@ -306,7 +306,8 @@ async function exec(): Promise<void> {
       added: [
         '**plugin-serlo-injection**. Added injection plugin for serlo.org content',
         '**plugin-table**. You can now create tables in markdown.'
-      ]
+      ],
+      fixed: ['**plugin-text**. Split plugin when pasting multiple blocks']
     }
   ])