+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - **plugin-table**. You can now create tables in markdown.
 
+### Changed
+
+- Optimize bundle size
+
 ## [0.7.0](https://github.com/edtr-io/edtr-io/compare/v0.6.2..v0.7.0) - July 6, 2019
 
 ### Breaking Changes