+++ CHANGELOG.md
@@ -14,6 +14,10 @@ All notable changes to this project will be documented in this file.
 - Optimize bundle size
 - **plugin-rows**. add-button is now always visible under the focused document
 
+### Fixed
+
+- **plugin-text**. Split plugin when pasting multiple blocks
+
 ## [0.7.0](https://github.com/edtr-io/edtr-io/compare/v0.6.2..v0.7.0) - July 6, 2019
 
 ### Breaking Changes